backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
altgraph @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
bitarray==2.9.2
blinker==1.7.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.0
ckzg==2.0.1
click==8.1.7
cloudinary==1.41.0
cryptography==43.0.1
cytoolz==1.0.0
distlib==0.3.8
Django==4.2.16
django-cors==0.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
ecdsa==0.19.0
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==5.0.0
eth-utils==5.0.0
eth_abi==5.1.0
factory_boy==3.3.1
Faker==30.3.0
filelock==3.13.4
Flask==3.0.0
Flask-Admin==1.6.1
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
future @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
git-filter-repo==2.45.0
gunicorn==23.0.0
hexbytes==1.2.1
idna==3.10
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
macholib @ file:///AppleInternal/Library/BuildRoots/860631e9-c1c5-11ee-98ee-b6ef2fd8d87b/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==2.1.3
multidict==6.1.0
mysql-connector-python==8.0.33
packaging==24.1
parsimonious==0.10.0
pipenv==2023.12.1
platformdirs==4.2.1
propcache==0.2.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
pyunormalize==16.0.0
regex==2024.9.11
requests==2.32.3
rlp==4.0.1
rsa==4.9
semantic-version==2.10.0
six==1.16.0
SQLAlchemy==2.0.29
sqlparse==0.5.1
stripe==11.1.0
toolz==1.0.0
types-requests==2.32.0.20240914
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.26.0
web3==7.3.0
websockets==13.1
Werkzeug==3.0.1
whitenoise==6.7.0
WTForms==3.1.2
yarl==1.14.0
zipp==3.17.0
